+++ src/components/gridColumn/index.jsx
@@ -128,7 +128,7 @@ GridColumn.propTypes = {
   lg: PropTypes.oneOf([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]),
   lgShift: PropTypes.oneOf([0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11]),
   includeGutter: PropTypes.bool,
-  gutterPosition: PropTypes.oneOf(["before, after"]),
+  gutterPosition: PropTypes.oneOf(["before", "after"]),
 };
 
 GridColumn.defaultProps = {