+++ src/components/multiStep/index.jsx
@@ -1,0 +1,2 @@
+export MultiStep from "./multiStep";
+export MultiStepWrapper from "./multiStepWrapper";