+++ src/components/text/textAccent.jsx
@@ -10,9 +10,9 @@ const styles = Object.assign({}, {
   marginTop: 0,
 }, textAccent());
 
-const TextAccent = ({ children, style }) => (
-  <p style={[styles, style]}>
-    {children}
+const TextAccent = (props) => (
+  <p {...props} style={[styles, props.style]}>
+    {props.children}
   </p>
 );