+++ CHANGELOG.md
@@ -2,6 +2,39 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="4.20.1"></a>
+## [4.20.1](https://github.com/lonelyplanet/backpack-ui/compare/v4.20.0...v4.20.1) (2017-05-25)
+
+
+### Bug Fixes
+
+* Add media query to disable motion in Toast ([#301](https://github.com/lonelyplanet/backpack-ui/issues/301)) ([5d6df27](https://github.com/lonelyplanet/backpack-ui/commit/5d6df27))
+* Delete unused watchLaterVideos from stories ([#283](https://github.com/lonelyplanet/backpack-ui/issues/283)) ([5279f7c](https://github.com/lonelyplanet/backpack-ui/commit/5279f7c))
+* Make bgDark 96% black ([#302](https://github.com/lonelyplanet/backpack-ui/issues/302)) ([7e5f436](https://github.com/lonelyplanet/backpack-ui/commit/7e5f436)), closes [#297](https://github.com/lonelyplanet/backpack-ui/issues/297)
+* Set width of WatchLaterModal to 630px ([#277](https://github.com/lonelyplanet/backpack-ui/issues/277)) ([1e16dda](https://github.com/lonelyplanet/backpack-ui/commit/1e16dda))
+
+
+
+<a name="4.20.0"></a>
+# [4.20.0](https://github.com/lonelyplanet/backpack-ui/compare/v4.19.1...v4.20.0) (2017-05-25)
+
+
+### Bug Fixes
+
+* Add "gray" to Button color prop validation ([#313](https://github.com/lonelyplanet/backpack-ui/issues/313)) ([c7117c2](https://github.com/lonelyplanet/backpack-ui/commit/c7117c2))
+* Add link to account help in social auth component ([#310](https://github.com/lonelyplanet/backpack-ui/issues/310)) ([5998b36](https://github.com/lonelyplanet/backpack-ui/commit/5998b36))
+* Set default href and role in Button to null ([#312](https://github.com/lonelyplanet/backpack-ui/issues/312)) ([83506a3](https://github.com/lonelyplanet/backpack-ui/commit/83506a3))
+* Spread props on text components ([#305](https://github.com/lonelyplanet/backpack-ui/issues/305)) ([71e64f1](https://github.com/lonelyplanet/backpack-ui/commit/71e64f1))
+* Update Heading with className and ID props ([#306](https://github.com/lonelyplanet/backpack-ui/issues/306)) ([23f43c2](https://github.com/lonelyplanet/backpack-ui/commit/23f43c2))
+
+
+### Features
+
+* Create Dialog component ([#307](https://github.com/lonelyplanet/backpack-ui/issues/307)) ([6995877](https://github.com/lonelyplanet/backpack-ui/commit/6995877))
+* Create Multi-step component ([#309](https://github.com/lonelyplanet/backpack-ui/issues/309)) ([aa4532b](https://github.com/lonelyplanet/backpack-ui/commit/aa4532b))
+
+
+
 <a name="4.19.1"></a>
 ## [4.19.1](https://github.com/lonelyplanet/backpack-ui/compare/v4.19.0...v4.19.1) (2017-05-12)